first

+++ exp.h
@@ -3,7 +3,7 @@
 
 #include "operations.h"
 
-Matrix* matrix_power(const Matrix* A, int power, Matrix* res);
+Matrix* matrix_power(const Matrix* A, size_t power, Matrix* res);
 
 Matrix* matrix_exponential(const Matrix *A, double eps);
